requirements.txt
yfinance>=0.2.40
gspread>=6.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
feedparser>=6.0.11
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
